Update README.md
This is a database app created with Microsoft Access (a database management system) to manage teachers schedules.  It provides a user-friendly interface.  The individual teachers accounts can be updated to optimize information management and ensure better analysis.The relationships between tables, input,  queries, forms, and reports have been defined in the images provided. That is the User case diagram, Sequence diagram, Activity diagram, and the Entity diagram.